requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
pydantic==2.6.4
httpx==0.27.0
python-dotenv==1.0.1
jinja2==3.1.3
pydantic-settings==2.2.1
dnspython==2.7.0
email-validator==2.2.0
pycryptodome==3.23.0
python-jose==3.4.0
python-multipart==0.0.20
psycopg[binary]==3.2.1